Add villain lair and battle methods to Game

+++ src/Invictus.java
@@ -13,7 +13,7 @@ public class Invictus extends Villain {
 	/**
 	 * Damage dealt by a villain to a hero's health when it wins the game.
 	 */
-	private static final double DAMAGE_DEALT = 50;
+	private static final int DAMAGE_DEALT = 50;
 	
 	/**
 	 * Array of games the villain could choose to play.
